backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart
opencv-python>=4.9.0
ultralytics>=8.1.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
